Change font to @fontsource-variable/instrument-sans
