+++ package.json
@@ -75,7 +75,7 @@
     "@babel/preset-env": "^7.22.20",
     "@babel/preset-react": "^7.22.15",
     "@babel/preset-typescript": "^7.23.0",
-    "@fontsource-variable/figtree": "^5.0.18",
+    "@fontsource-variable/instrument-sans": "^5.0.3",
     "@mapbox/rehype-prism": "^0.8.0",
     "@mdx-js/loader": "^2.3.0",
     "@mdx-js/react": "^2.3.0",
