+++ src/app/guide/installation/installation.mdx
@@ -98,12 +98,12 @@ Some fonts in Web applications do not render best on the _Windows_ platform, or 
 
 > You don't need to modify any CSS files or declare fonts.
 
-## Install `@fontsource-variable/inter`:
+## Install `@fontsource-variable/instrument-sans`:
 
-<Snippet>yarn add@fontsource-variable/figtree</Snippet>
+<Snippet>yarn add @fontsource-variable/instrument-sans</Snippet>
 
-## Add `Figtree Variable` to your project:
+## Add `Instrument Sans Variable` to your project:
 
 <Code block my={0} width="100%">
-  <span className="token keyword module">import</span> <span className="token string">'@fontsource-variable/figtree'</span>
+  <span className="token keyword module">import</span> <span className="token string">'@fontsource-variable/instrument-sans'</span>
 </Code>
