+++ src/components/quick-bar/index.ts
@@ -2,6 +2,7 @@ import { default as QuickBar } from './quick-bar';
 export { default as QuickAction } from './quick-action';
 export { default as QuickBarLayout } from './quick-bar-layout';
 
-export type { QuickActionProps, QuickBarLayoutProps, QuickBarProps } from './share';
+export type { QuickActionProps, QuickBarLayoutProps } from './share';
+export type { QuickBarProps } from './quick-bar';
 
 export default QuickBar;
