+++ src/components/quick-bar/share.ts
@@ -4,12 +4,6 @@ import { HTMLAttributeAnchorTarget } from 'react';
 export interface QuickBarLayoutProps {
   animationTime?: number;
 }
-interface NativeQuickBarProps {
-  header?: React.ReactNode;
-}
-
-type NativeAttrs = Omit<React.HTMLAttributes<HTMLDivElement>, keyof NativeQuickBarProps>;
-export type QuickBarProps = NativeQuickBarProps & NativeAttrs;
 
 export interface QuickActionProps {
   tooltip?: string | React.ReactNode;
