+++ src/components/snippet/snippet.tsx
@@ -93,7 +93,7 @@ const SnippetComponent: React.FC<React.PropsWithChildren<SnippetProps>> = ({
           width: ${SCALES.width(1, 'initial')};
           height: ${SCALES.height(1, 'auto')};
           padding: ${SCALES.pt(0.667)} ${SCALES.pr(2.667)} ${SCALES.pb(0.667)} ${SCALES.pl(0.667)};
-          margin: ${SCALES.mt(1)} ${SCALES.mr(0)} ${SCALES.mb(1)} ${SCALES.ml(0)};
+          margin: ${SCALES.mt(0)} ${SCALES.mr(0)} ${SCALES.mb(0)} ${SCALES.ml(0)};
         }
 
         pre {
